10/4/23 - Successfully animated dice roll in DiceLogic (Updated MainActivity.kt and DiceLogic.kt)
